Fix and reorganize tests
