+++ jest.config.js
@@ -12,7 +12,7 @@ module.exports = {
         '^.+\\.ts?$': 'ts-jest',
     },
     moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
-    modulePathIgnorePatterns: ['mocks'],
+    modulePathIgnorePatterns: ['utils'],
     setupFilesAfterEnv: [
         "@testing-library/jest-native/extend-expect",
         "./setup-tests.js"
