+++ src/__tests__/Variable.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Literal, Variable as VariableModel } from 'wollok-ts/dist/model'
-import { renderWithTheme } from '../../../../../utils/test-helpers'
-import { Variable } from '../Variable'
+import { renderWithTheme } from './utils/test-helpers'
+import { Variable } from '../components/Body/sentences/Variable'
 
 function renderVariable(variable: VariableModel) {
 	const { UNSAFE_queryByProps, getByText } = renderWithTheme(
