+++ src/__tests__/utils/test-helpers.tsx
@@ -1,0 +1,14 @@
+import { NavigationContainer } from '@react-navigation/native'
+import { render } from '@testing-library/react-native'
+import React from 'react'
+import { theme } from '../../theme'
+import ProjectProviderMock from './ProjectProviderMock'
+import { OneOrMany } from '../../utils/type-helpers'
+
+export function renderWithTheme(children: OneOrMany<JSX.Element>) {
+	return render(
+		<ProjectProviderMock>
+			<NavigationContainer theme={theme}>{children}</NavigationContainer>
+		</ProjectProviderMock>,
+	)
+}
