+++ src/components/Body/sentences/Assignment.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { Assignment as AssignmentModel } from 'wollok-ts/dist/model'
-import { ReferenceSegment } from '../../../expressions/expression-segment'
-import { display } from '../../../expressions/ExpressionDisplay'
-import { ProblemReporterButton } from '../../../problems/ProblemReporterButton'
-import { Row } from '../../Row'
+import { ReferenceSegment } from '../../expressions/expression-segment'
+import { display } from '../../expressions/ExpressionDisplay'
+import { ProblemReporterButton } from '../../problems/ProblemReporterButton'
+import { Row } from '../../ui/Row'
 import { ReferenceExpression } from './ReferenceExpression'
 
 export const Assignment = ({ assignment }: { assignment: AssignmentModel }) => {
