+++ src/components/Body/sentences/Send.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Send as SendModel } from 'wollok-ts/dist/model'
-import { ExpressionDisplay } from '../../../expressions/ExpressionDisplay'
-import { ProblemReporterButton } from '../../../problems/ProblemReporterButton'
-import { Row } from '../../Row'
+import { ExpressionDisplay } from '../../expressions/ExpressionDisplay'
+import { ProblemReporterButton } from '../../problems/ProblemReporterButton'
+import { Row } from '../../ui/Row'
 
 export const Send = ({ send }: { send: SendModel }) => {
 	return (
