+++ src/components/Body/sentences/VisualSentence.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Text } from 'react-native-paper'
 import { Sentence } from 'wollok-ts/dist/model'
-import { wTranslate } from '../../../../utils/translation-helpers'
-import { Row } from '../../Row'
+import { wTranslate } from '../../../utils/translation-helpers'
+import { Row } from '../../ui/Row'
 import { Assignment } from './Assignment'
 import { Return } from './Return'
 import { Send } from './Send'
